voucher-detail-page

+++ src/components/VoucherListRow.jsx
@@ -1,9 +1,11 @@
 import React, { useState } from "react";
-import { HiOutlineTrash } from "react-icons/hi";
+import { HiOutlinePencil, HiOutlineTrash } from "react-icons/hi";
 import ShowDate from "./showDate";
 import { lineSpinner } from "ldrs";
 import toast from "react-hot-toast";
 import { useSWRConfig } from "swr";
+import { Link } from "react-router-dom";
+import { TbListDetails } from "react-icons/tb";
 lineSpinner.register();
 
 const VoucherListRow = ({
@@ -14,17 +16,19 @@ const VoucherListRow = ({
   const { mutate } = useSWRConfig();
 
   const handleDelete = async () => {
-    setIsDeleting(true);
+    if (confirm("Are you sure you want to delete this voucher?")) {
+      setIsDeleting(true);
 
-    await fetch(import.meta.env.VITE_API_URL + `/vouchers/${id}`, {
-      method: "DELETE",
-    });
+      await fetch(import.meta.env.VITE_API_URL + `/vouchers/${id}`, {
+        method: "DELETE",
+      });
 
-    toast.success("Product deleted successfully");
+      toast.success("Product deleted successfully");
 
-    mutate(import.meta.env.VITE_API_URL + "/vouchers");
+      mutate(import.meta.env.VITE_API_URL + "/vouchers");
 
-    setIsDeleting(false);
+      setIsDeleting(false);
+    }
   };
   return (
     <tr className="odd:bg-white  even:bg-gray-50  border-b border-gray-200">
@@ -37,25 +41,37 @@ const VoucherListRow = ({
       <td className="px-6 py-4">{customer_name}</td>
       <td className="px-6 py-4">{customer_email}</td>
       <td className="px-6 py-4 text-end">
-        <ShowDate timestamp={sale_date} />
+        <ShowDate timestamp={sale_date} hasTime={false} />
       </td>
       <td className="px-6 py-4 text-end">
-        <button
-          onClick={handleDelete}
-          type="button"
-          className="size-8 flex justify-end items-center text-sm font-medium text-red-500 hover:text-red-700"
+        <div
+          className="inline-flex shadow-xs border border-gray-500"
+          role="group"
         >
-          {isDeleting ? (
-            <l-line-spinner
-              size="18"
-              stroke="1"
-              speed="1"
-              color="red"
-            ></l-line-spinner>
-          ) : (
-            <HiOutlineTrash />
-          )}
-        </button>
+          <Link
+            to={`/voucher/detail/${id}`}
+            type="button"
+            className="size-8 flex justify-center items-center text-sm font-medium text-sky-500 bg-white hover:bg-gray-200 hover:text-sky-700"
+          >
+            <TbListDetails />
+          </Link>
+          <button
+            onClick={handleDelete}
+            type="button"
+            className="size-8 flex justify-center items-center text-sm font-medium text-red-500 bg-white hover:bg-gray-200 hover:text-red-700"
+          >
+            {isDeleting ? (
+              <l-line-spinner
+                size="18"
+                stroke="1"
+                speed="1"
+                color="red"
+              ></l-line-spinner>
+            ) : (
+              <HiOutlineTrash />
+            )}
+          </button>
+        </div>
       </td>
     </tr>
   );
